cassandra_client/requirements.txt
cassandra-driver==3.13.0
gunicorn==19.7.1
Flask==0.12.2
Flask-RESTful==0.3.6
